Avoid failure to instantiate an OperationFailure if details is not a dict Unit test BulkWriteError class Document the details parameter on OperationFailure init Document the results parameter on BulkWriteError init Avoid failure on tearDown in case of a unittest.TestCase

+++ test/__init__.py
@@ -690,12 +690,13 @@ def setup():
 
 def teardown():
     c = client_context.client
-    c.drop_database("pymongo-pooling-tests")
-    c.drop_database("pymongo_test")
-    c.drop_database("pymongo_test1")
-    c.drop_database("pymongo_test2")
-    c.drop_database("pymongo_test_mike")
-    c.drop_database("pymongo_test_bernie")
+    if c:
+        c.drop_database("pymongo-pooling-tests")
+        c.drop_database("pymongo_test")
+        c.drop_database("pymongo_test1")
+        c.drop_database("pymongo_test2")
+        c.drop_database("pymongo_test_mike")
+        c.drop_database("pymongo_test_bernie")
 
 
 class PymongoTestRunner(unittest.TextTestRunner):
